+++ hal-baremetal/src/hardware.c
@@ -164,13 +164,13 @@ void hardware_rtc_set(const struct hardware_time *tm_hw) {
 	struct rtc_time tm_rtc;
 	struct tm tmbuf;
 
-	tm_rtc.tm_hour = tm_hw->hour;
-	tm_rtc.tm_min = tm_hw->minute;
-	tm_rtc.tm_sec = tm_hw->second;
-	tm_rtc.tm_mday = tm_hw->day;
+	tm_rtc.tm_hour = (int)tm_hw->hour;
+	tm_rtc.tm_min = (int)tm_hw->minute;
+	tm_rtc.tm_sec = (int)tm_hw->second;
+	tm_rtc.tm_mday = (int)tm_hw->day;
 	//tm_rtc.tm_wday = // TODO
-	tm_rtc.tm_mon = tm_hw->month;
-	tm_rtc.tm_year = tm_hw->year - 2000;	// RTC stores 2 digits only
+	tm_rtc.tm_mon = (int)tm_hw->month;
+	tm_rtc.tm_year = (int)tm_hw->year - 2000;	// RTC stores 2 digits only
 
 	if (mcp7941x_start(0x00) != MCP7941X_ERROR) {
 		mcp7941x_set_date_time(&tm_rtc);
